requirements.txt
zeep~=4.0.0
selenium~=3.141.0
Pillow~=8.2.0
Telethon~=1.21.1